% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.classify_b}
\alias{.classify_b}
\title{Classify species based on a single variable and numeric thresholds (Criterion B)}
\usage{
.classify_b(df, thresholds, var_th)
}
\arguments{
\item{df}{A data frame containing the variable to classify.}

\item{thresholds}{A numeric vector of length 3, defining the thresholds for:
\itemize{
\item \code{CR}: values < thresholds \verb{[1]}
\item \code{EN}: values < thresholds \verb{[2]}
\item \code{VU}: values < thresholds \verb{[3]}
}}

\item{var_th}{A character string indicating the name of the column in \code{df} to classify.}
}
\value{
A character vector with classification results:
\itemize{
\item \code{"CR"} = Critically Endangered
\item \code{"EN"} = Endangered
\item \code{"VU"} = Vulnerable
\item \code{NA} if no thresholds are met
}
}
\description{
Internal function used to assign a Red List category (\code{CR}, \code{EN}, \code{VU}) based on a
continuous numeric variable and threshold values.
}
\details{
Typically used to classify geographic range measures such as:
\itemize{
\item Extent of Occurrence (EoO, \code{B1_eoo})
\item Area of Occupancy (AoO, \code{B2_aoo})
}
}
