% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.evaluate_ba}
\alias{.evaluate_ba}
\title{Evaluate criterion B(a): small population size}
\usage{
.evaluate_ba(df)
}
\arguments{
\item{df}{A data frame that must include the variable "B_AND_a".}
}
\value{
A character vector with values: "CR", "EN", "VU", or NA.
}
\description{
Assigns a conservation category (CR, EN, VU) based on the number of individuals
in a population.
}
