% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.evaluate_bb}
\alias{.evaluate_bb}
\title{Evaluate criterion B(b): continuing decline}
\usage{
.evaluate_bb(df)
}
\arguments{
\item{df}{A data frame that must include the variables "B_AND_b_i" to "B_AND_b_v".}
}
\value{
A string of the form "b(i,ii,...)" if any sub-criterion is met, or NA
if none apply.
}
\description{
Checks which sub-criteria under criterion B(b) are triggered (i.e., have value
1 or are TRUE), and returns a label indicating which ones.
}
