% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.evaluate_bc}
\alias{.evaluate_bc}
\title{Evaluate criterion B(c): extreme fluctuations}
\usage{
.evaluate_bc(df)
}
\arguments{
\item{df}{A data frame that must include the variables "B_AND_c_i" to "B_AND_c_iv".}
}
\value{
A string of the form "c(i,ii,...)" if any sub-criterion is met, or NA if none apply.
}
\description{
Checks which sub-criteria under criterion B(c) are triggered (i.e., have value 1), and returns a label indicating which ones.
}
