% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.evaluate_cmat}
\alias{.evaluate_cmat}
\title{Evaluate Criterion C: number of mature individuals
\subsection{Documentar !!!!!}{
}}
\usage{
.evaluate_cmat(df)
}
\description{
Evaluate Criterion C: number of mature individuals
\subsection{Documentar !!!!!}{
}
}
