% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.evaluate_bb}
\alias{.evaluate_bb}
\title{Evaluate criterion B(b): continuing decline}
\usage{
.evaluate_bb(df)
}
\arguments{
\item{df}{A data frame containing variables \code{B_AND_b_i} to \code{B_AND_b_v} (or lowercased).}
}
\value{
A character vector with the same number of rows as \code{df}. Each value is:
\itemize{
\item \code{"b(i,...)"} if subcriteria are active
\item \code{NA} if none apply
}
}
\description{
Identifies which subcriteria under Criterion B(b) are active (i.e., have value
1 or are TRUE)), and returns a label indicating the codes, such as \code{"b(i,ii,iv)"}.
}
