% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.evaluate_bc}
\alias{.evaluate_bc}
\title{Evaluate Criterion B(c): extreme fluctuations}
\usage{
.evaluate_bc(df)
}
\arguments{
\item{df}{A data frame containing variables \code{B_AND_c_i} to \code{B_AND_c_iv} (or lowercased).}
}
\value{
A character vector with the same number of rows as \code{df}. Each value is:
\itemize{
\item \code{"c(i,...)"} if subcriteria are active
\item \code{NA} if none apply
}
}
\description{
Identifies which subcriteria under Criterion B(c) are active (i.e., have value
1 or are TRUE)), and returns a label indicating the codes, such as \code{"c(ii,iv)"}.
}
